# openapi.yaml
openapi: 3.0.3
info:
  title: TCSS 460 Group Project API
  description: |
    **Team:** Group 4
    **Deployed URL:** https://group-project-backend-group-4.onrender.com

    A RESTful API for a move and TV review community.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://group-project-backend-group-4.onrender.com
    description: Production deployment
...
paths:
...
  /issues/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: The ID of the issue.
    get:
      tags: [Issues]
      summary: Get Issue by ID
      description: Returns a single issue by its ID.
      responses:
        '200':
          description: Issue retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  issue:
                    $ref: '#/components/schemas/Issue'
        '400':
          description: Invalid ID parameter.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Validation failed
                  details:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: string
                          example: id
                        message:
                          type: string
                          example: Expected number, received string
        '404':
          description: Issue not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Issue not found
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Internal Server Error
    put:
      tags: [Issues]
      summary: Update Issue
      description: Updates an existing issue's details or status.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                body:
                  type: string
                contact:
                  type: string
                status:
                  $ref: '#/components/schemas/IssueStatus'
      responses:
        '200':
          description: Issue updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Issue updated"
                  newIssue:
                    $ref: '#/components/schemas/Issue'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Validation failed
                  details:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: string
                        message:
                          type: string
        '404':
          description: Issue not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Issue not found
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Internal Server Error
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Issue:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "Broken search"
        body:
          type: string
          example: "Search fails for movies with special characters."
        contact:
          type: string
          example: "user@example.com"
        status:
          $ref: '#/components/schemas/IssueStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    IssueStatus:
      type: string
      enum: [OPEN, IN_PROGRESS, RESOLVED, WONT_FIX]
      example: OPEN